完成shop-service

+++ spike-system-api/src/main/java/com/ogic/spikesystemapi/entity/ShopEntity.java
@@ -32,5 +32,5 @@ public class ShopEntity {
     /**
      * 店铺余额
      */
-    public Integer money;
+    public Double money;
 }
